address review

+++ src/server/server_state.h
@@ -138,7 +138,7 @@ class ServerState {  // public struct - to allow initialization.
     std::valarray<uint64_t> tx_width_freq_arr, squash_width_freq_arr;
 
     // Memory size of stored commands during multi-exec in connections
-    uint64_t stored_cmd_bytes = 0;
+    size_t stored_cmd_bytes = 0;
   };
 
   // Unsafe version.
